Initial commit
HTTP Signatures can be signed and verified with RSA and HMAC
There are traits to implement to add and Authorization header
to a request, and this is implemented for Reqwest and Hyper.

There has been no effort for server-side code so far, but
a request guard for Rocket is planned.