Add checks on headers
For creators, headers must not be empty. This is because if any header
is added to the signing string, it is also present in the headers field.

For verifiers, assume that a lack of headers implies a single header with
the name 'date'. This is in compliance with the specification.